Fixed int overflow in EncogEGBFile
git-svn-id: http://encog-java.googlecode.com/svn/trunk/encog-core@3510 f90f6e9a-ac51-0410-b353-d1b83c6f6923

+++ src/main/java/org/encog/ml/data/buffer/EncogEGBFile.java
@@ -252,8 +252,8 @@ public class EncogEGBFile {
 	 *            The row to calculate for.
 	 * @return The index.
 	 */
-	private int calculateIndex(final int row) {
-		return EncogEGBFile.HEADER_SIZE + (row * this.recordSize);
+	private long calculateIndex(final long row) {
+		return (long)EncogEGBFile.HEADER_SIZE + (row * (long)this.recordSize);
 	}
 
 	/**
